Quick fixes (see description for list of fixes)
-PhotonView id conflict between BossCage and Manager.  Both were made 0
in different scenes. Changed BossCage to 900.
-MonsterClient was trying to instantiate the jungle camps as scene
objects on every client. Added in a check for IsMasterClient() to the
NetworkManager and checked it before instantiating camps.
- Fixed RPC resolution error with player movement: Enums are cast as
ints when sent in RPCs, so I changed the calls to intentionally cast it
to int (for the sake of its own client) and then accept the int and cast
it back to a PlayerInput.TargetType in the RPC declaration.